vendor: camera_config: s5k4e8: Declare as BACK camera
The BACK_AUX camera is auxilliary to the BACK camera, meaning
that it is used for functions like MONO picture or for 3DToF,
but not usable as a camera by itself.

Set this camera as BACK in order for it to be selectable and
usable for all of the regular video and shapshot usecases.